+++ src/community/components/card/card.stories.tsx
@@ -1,12 +1,12 @@
 import { Meta, StoryFn, StoryObj } from '@storybook/react';
 
 import { Heading } from '../../../tedi/components/base/typography/heading/heading';
+import { Collapse } from '../../../tedi/components/buttons/collapse/collapse';
 import { Col, Row } from '../../../tedi/components/layout/grid';
 import { VerticalSpacing } from '../../../tedi/components/layout/vertical-spacing';
 import Separator from '../../../tedi/components/misc/separator/separator';
 import { StretchContent } from '../../../tedi/components/misc/stretch-content/stretch-content';
 import { CardsExample } from '../../../tedi/components/misc/stretch-content/stretch-content.stories';
-import Collapse from '../collapse/collapse';
 import Icon from '../icon/icon';
 import { Card, CardProps } from './card';
 import CardContent, { CardContentPadding, CardContentProps } from './card-content/card-content';