+++ src/community/components/layout/header/components/header-role/header-role.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
 import { Text } from '../../../../../../tedi/components/base/typography/text/text';
+import { Collapse } from '../../../../../../tedi/components/buttons/collapse/collapse';
 import ScrollFade from '../../../../../../tedi/components/misc/scroll-fade/scroll-fade';
 import { useLabels } from '../../../../../../tedi/providers/label-provider';
 import { Card, CardContent } from '../../../../card';
-import Collapse from '../../../../collapse/collapse';
 import Tag from '../../../../tag/tag';
 import { HeaderDropdown } from '../header-dropdown/header-dropdown';
 import styles from './header-role.module.scss';